server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
pydantic[email]>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
bcrypt>=4.1.0
